+++ src/app/modules/home/home.component.ts
@@ -4,8 +4,9 @@ import { setUrlParam } from '../../shared/utils';
 import { RasTokenService } from '../../shared/ras-token/ras-token.service';
 import { StepperComponent, Step } from '../stepper/stepper.component';
 import { CreateCohortMode } from '../../shared/cohort/cohort.service';
-import { FhirBackendService } from '../../shared/fhir-backend/fhir-backend.service';
+import {ConnectionStatus, FhirBackendService} from '../../shared/fhir-backend/fhir-backend.service';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
+import {Oauth2TokenService} from "../../shared/oauth2-token/oauth2-token.service";
 
 @Component({
   selector: 'app-home',
@@ -19,7 +20,8 @@ export class HomeComponent implements AfterViewInit {
   constructor(
     public rasToken: RasTokenService,
     public fhirBackend: FhirBackendService,
-    private liveAnnouncer: LiveAnnouncer
+    private liveAnnouncer: LiveAnnouncer,
+    public oauth2Token: Oauth2TokenService
   ) {}
 
   openChangelog(): void {
@@ -56,13 +58,35 @@ export class HomeComponent implements AfterViewInit {
     );
   }
 
+  /**
+   * Initiate login through OAuth2.
+   */
+  onOauth2Login(): void {
+    this.oauth2Token.login(this.fhirBackend.serviceBaseUrl);
+  }
+
+  private returnToSettingsPage(): void {
+    this.stepperComponent.stepper.selectedIndex = Step.SETTINGS;
+    this.stepperComponent.selectAnActionComponent?.createCohortMode.setValue(
+      CreateCohortMode.UNSELECTED
+    );
+    this.liveAnnouncer.announce('Logged out. Returning to settings page.');
+  }
+
+  /**
+   * Log out of a server connected through OAuth2.
+   */
+  onOauth2Logout(): void {
+    this.oauth2Token.logout();
+    this.returnToSettingsPage();
+    // Show "Authorization required" error message below server input.
+    this.fhirBackend.initialized.next(ConnectionStatus.Oauth2Required);
+    this.stepperComponent.settingsPageComponent.settingsFormGroup.get('serviceBaseUrl').updateValueAndValidity();
+  }
+
   onRasLogout(): void {
     this.rasToken.logout().then(() => {
-      this.stepperComponent.stepper.selectedIndex = Step.SETTINGS;
-      this.stepperComponent.selectAnActionComponent.createCohortMode.setValue(
-        CreateCohortMode.UNSELECTED
-      );
-      this.liveAnnouncer.announce('Logged out. Returning to settings page.');
+      this.returnToSettingsPage();
     });
   }